实现了简单的注册接口

+++ member/src/main/java/com/train/entry/Member.java
@@ -4,6 +4,7 @@ import lombok.Data;
 
 @Data
 public class Member {
-    private Integer id;
+
+    private Long id;
     private String mobile;
 }
